Move project boilerplate to a util file
